feat: allow key signatures at start of volpiano strings
Opening clefs may now be immediately followed by some number of accidentals (currently supported flats on B lines/spaces are included, as are the future characters for E flats, F sharps, and C sharps). If correctly encoded, this opening material is stored during volpiano syllabification and prepended to the aligned text and volpiano.